Added expandable/contractable sidebar. Moved datasets code from sidebar into datasets component.
