+++ src/app/layout/datasets/datasets.component.scss
@@ -1,0 +1,56 @@
+.card-body {
+    padding-top: 10px;
+    padding-left: 10px;
+    padding-right: 10px;
+    .row {
+        margin-bottom: 10px;
+    }
+    li {
+        list-style-type: none;
+    }
+    .submenu-header {
+        padding: 4px;
+        cursor: pointer;
+        background: #eee;
+    }
+}
+
+// Datasets CSS below
+.btn {
+    cursor: pointer;
+}
+
+// Fix issue with datepicker z-index
+::ng-deep ngb-datepicker.dropdown-menu.ng-star-inserted {
+    //left: 0px !important;
+    z-index: 999999;
+}
+
+.records-spinner {
+    position: relative;
+    top: 50%;
+    left: 50%;
+    -ms-transform: translate(-50%,-50%); /* IE 9 */
+    -webkit-transform: translate(-50%,-50%); /* Safari */
+    transform: translate(-50%,-50%); /* Standard syntax */
+    width: 70px;
+    height: 70px;
+    > div {
+        width: 18px;
+        height: 18px;
+        background-color: white;
+        border-radius: 100%;
+        display: inline-block;
+        -webkit-animation: sk-bouncedelay 1.4s infinite ease-in-out both;
+        animation: sk-bouncedelay 1.4s infinite ease-in-out both;
+    }
+    .bounce1 {
+        -webkit-animation-delay: -0.32s;
+        animation-delay: -0.32s;
+    }
+
+    .bounce2 {
+        -webkit-animation-delay: -0.16s;
+        animation-delay: -0.16s;
+    }
+}
